+++ src/main/java/es/codeurjc/web/service/ValoracionService.java
@@ -1,18 +1,24 @@
 package es.codeurjc.web.service;
 
+import java.util.Optional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import es.codeurjc.web.model.Valoracion;
-import es.codeurjc.web.model.Usuario;
 import es.codeurjc.web.model.Consejo;
+import es.codeurjc.web.model.Usuario;
+import es.codeurjc.web.model.Valoracion;
 import es.codeurjc.web.repository.ValoracionRepository;
-import java.util.Optional;
 
 @Service
 public class ValoracionService {
 
     @Autowired
     private ValoracionRepository valoracionRepository;
+
+    @Autowired
+    private UsuarioService usuarioService;
+
+    @Autowired
+    private ConsejoService consejoService;
 
     public void save(Valoracion valoracion) {
         valoracionRepository.save(valoracion);
@@ -26,7 +32,32 @@ public class ValoracionService {
         return valoracionRepository.findByAuthorAndConsejo(author, consejo);
     }
 
-    public void deleteById(Long id) {
-        valoracionRepository.deleteById(id);
+    public void createReview(Long consejoId, String userEmail, String title, int score, String comment) {
+        Usuario user = usuarioService.findByEmail(userEmail).orElseThrow();
+        Consejo consejo = consejoService.findById(consejoId).orElseThrow();
+        Valoracion v = new Valoracion(user, consejo, score, title, comment);
+        valoracionRepository.save(v);
+    }
+
+    public boolean updateReview(Long id, String userEmail, String title, int score, String comment) {
+        Optional<Valoracion> vOpcional = valoracionRepository.findById(id);
+        if (vOpcional.isPresent() && vOpcional.get().getAuthor().getEmail().equals(userEmail)) {
+            Valoracion v = vOpcional.get();
+            v.setTitle(title);
+            v.setScore(score);
+            v.setComment(comment);
+            valoracionRepository.save(v);
+            return true;
+        }
+        return false;
+    }
+
+    public boolean deleteReview(Long id, String userEmail) {
+        Optional<Valoracion> vOpcional = valoracionRepository.findById(id);
+        if (vOpcional.isPresent() && vOpcional.get().getAuthor().getEmail().equals(userEmail)) {
+            valoracionRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
